refactoring, extract the remove-components code from the Consumer class and move to a new file src/consumer/component-ops/remove-components.js
